third_party/blink/web_tests/external/wpt/css/css-inline/baseline-source/baseline-source-last-textarea-002.tentative-expected.txt

This is a testharness.js-based test.
FAIL .target > * 1 assert_equals: 
<span data-offset-x="46"></span>
offsetLeft expected 46 but got 60
FAIL .target > * 2 assert_equals: 
<textarea data-offset-x="11" class="inner"></textarea>
offsetLeft expected 11 but got 10
FAIL .target > * 3 assert_equals: 
<span data-offset-x="46"></span>
offsetLeft expected 46 but got 10
FAIL .target > * 4 assert_equals: 
<textarea data-offset-x="11" class="inner">X X X X X</textarea>
offsetLeft expected 11 but got 20
FAIL .target > * 5 assert_equals: 
<span data-offset-x="46"></span>
offsetLeft expected 46 but got 60
FAIL .target > * 6 assert_equals: 
<textarea data-offset-x="11" class="inner" placeholder="X X X X X"></textarea>
offsetLeft expected 11 but got 10
FAIL .target > * 7 assert_equals: 
<span data-offset-x="46"></span>
offsetLeft expected 46 but got 25
FAIL .target > * 8 assert_equals: 
<textarea data-offset-x="11" class="inner" style="font-size: 100px;"></textarea>
offsetLeft expected 11 but got 10
FAIL .target > * 9 assert_equals: 
<span data-offset-x="46"></span>
offsetLeft expected 46 but got 10
FAIL .target > * 10 assert_equals: 
<textarea data-offset-x="11" class="inner" style="font-size: 100px;">X X X X X</textarea>
offsetLeft expected 11 but got 20
FAIL .target > * 11 assert_equals: 
<span data-offset-x="46"></span>
offsetLeft expected 46 but got 10
FAIL .target > * 12 assert_equals: 
<textarea data-offset-x="11" class="inner" style="font-size: 100px;" placeholder="X X X X X"></textarea>
offsetLeft expected 11 but got 10
Harness: the test ran to completion.
